requirements.txt
Django==1.8.4
Pillow==2.9.0
Unidecode==0.4.18
Willow==0.2.1
argparse==1.2.1
beautifulsoup4==4.4.0
django-appconf==1.0.1
django-compressor==1.5
django-modelcluster==0.6.2
django-taggit==0.17.1
django-treebeard==3.0
djangorestframework==3.2.3
html5lib==0.999
pytz==2015.4
six==1.9.0
wagtail==1.1
wagtailsettings==0.4.1
wsgiref==0.1.2
